% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_handlers.R
\name{convert_path_data}
\alias{convert_path_data}
\title{Convert the data derived from a path on a sunburst sector to a data.frame.}
\usage{
convert_path_data(x, session, inputname)
}
\arguments{
\item{x}{The data obtained from the JavaScript Sunburst visualization for a moused-over path.
This should already have been transformed with \code{jsonlite::fromJSON} and be a list of
lists.}

\item{session, inputname}{shiny session and input-name for the element.  Not used.}
}
\value{
A data.frame with columns 'group', 'name', 'value' and 'percentage'. This contains a
row for each ring of the Sunburst diagram (up to the ring that the mouse is hovering over).
}
\description{
Convert the data derived from a path on a sunburst sector to a data.frame.
}
