% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{encode_color_overrides}
\alias{encode_color_overrides}
\title{Ensure color-override definitions contain hex-coded color strings}
\usage{
encode_color_overrides(x)
}
\arguments{
\item{x}{list of lists. Each entry must have a `color` entry, and may have a `name` or
`group` entry. The `color` may be a named R color, an \code{rgb()} value or a hex string.}
}
\value{
A list of lists. The output will be identical to the input, except that any `color`
  entry will have been converted to a hex string.
}
\description{
Ensure color-override definitions contain hex-coded color strings
}
