% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_handlers.R
\name{get_shiny_input_handler}
\alias{get_shiny_input_handler}
\title{Get the shiny input-handler code for use on the JS side of a shiny app}
\usage{
get_shiny_input_handler(inputId, type = c("path_data", "set_to_null"))
}
\arguments{
\item{inputId}{the shiny input ID into which any event-driven data will be injected.}

\item{type}{the type of data that is to be extracted from the Sunburst widget. Choices are:
\itemize{
\item "path_data" - extracts a data.frame of "group", "name", "value" and "percentage" values
associated with a sunburst path.
\item "set_to_null" - sets the value of \code{input[[inputId]]} to NULL.
}}
}
\value{
a \code{JS_EVAL} object that can be used as the mouseover_handler or mouseout_handler in
the sunburst() function.
}
\description{
Get the shiny input-handler code for use on the JS side of a shiny app
}
