% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mult}
\alias{mult}
\title{Temporary function - to ensure CI pipeline passes on a trivial package}
\usage{
mult(x, y)
}
\arguments{
\item{x, y}{Numbers}
}
\value{
The product of x and y
}
\description{
Temporary function - to ensure CI pipeline passes on a trivial package
}
