% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{parse_js_colors}
\alias{parse_js_colors}
\title{Converts the string \code{"rgb(10, 20, 30)"} to a hexstring for the corresponding color}
\usage{
parse_js_colors(x)
}
\arguments{
\item{x}{Character vector. The contents of this vector must be of the form
\code{"rgb(0, 100, 255)"}.}
}
\value{
A character vector of the same length as \code{x}. Containing hexstrings that define
the same colors.
}
\description{
When colors are sent from JavaScript to R, they are received as strings of the form
\code{"rgb(x, y, z)"}, rather than as calls to the R \code{rgb()} function. Hence the need to parse a
color string that can be used within R.
}
