% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Create a HTML sunburst plot for displaying admissions and retention data}
\usage{
sunburst(
  data,
  steps,
  palette = NULL,
  color_overrides = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Data-frame. This should have one row per observation (student).}

\item{steps}{Character. Subset of the columns of the data-frame. These correspond to the
rings of the sunburst chart. The first entry of steps is the inner-most ring of the chart.}

\item{palette}{Character. A vector of colors that determines the color of each sector in the
inner-most ring of the sunburst chart. The sectors are colored from largest to smallest.}

\item{color_overrides}{List of lists. Each sublist should have a 'color' entry (containing
any valid R color) and should have at least one of 'group' and 'name'. To specify a color for
a ring of the Sunburst chart, provide a 'group' entry. To specify a color for a named sector
of a ring, provide a 'name' entry. The 'group' will be a column name in \code{data} and the
'name' will be one of the levels within a column of \code{data}. Since a given name may be
present in different columns, you can specify both a group and a name to be more explicit.}

\item{width, height}{The initial size of the visualization}

\item{elementId}{Identifier for the HTML element into which the visualization will be added.}
}
\description{
Create a HTML sunburst plot for displaying admissions and retention data
}
