% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Create a HTML sunburst plot for displaying admissions and retention data}
\usage{
sunburst(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Data-frame.}

\item{width, height}{The initial size of the visualization}

\item{elementId}{Identifier for the HTML element into which the visualization will be added.}
}
\description{
Create a HTML sunburst plot for displaying admissions and retention data
}
